discovery-data/webhook-enrichment-sample/granite/requirements.txt
Flask
pyjwt
requests
